V0.43 More Shape Drawing and Collisions
Made it so when a 1 is read from the text file, a AABB with the exact
dimensions of a square in the grid is created. The square can be
collided with the ant's radius automatically. The ant should know which
direction to move in after a collision. There is a small issue of the
ant moving in the wall but some position correction will fix that.
Fixed a major problem in main where shapes were getting added on the
vector every frame so the longer the program ran, the slower it would
run because of unnecessary drawing and collision tests.
Disabled the drawing of the grid as it causes major lag. Uncomment to
turn back on