0.82 Food behaviour 99% done
when the ants radius collides with the foods, it will start moving
towards the food. when the food is picked up, the ant will move back to
the ant hill to drop the food off and it will choose another random
direction to move in.
the only problem ive encountered is when the ants are not following each
other. if an ant collides with food, for some reason, only the first ant
will pick up the food. so the food will warp to that ant. not sure why
this is happening because the food should go to the ant that collided
with it