AntEater Commit v0.2 - incomplete
Almost done, just here for the sake of backing up since I don't have my
portable with me and this PC is temperamental at best. Not complete,
expect the complete one later tonight with collisions.

Anteater class and source file included. Mostly same as ant, created a
vector to store the lone anteater so it can be expanded to include
multiple if we want to go crazy with the AI driven version.  Either way,
it's its own mind now.