AntEater Commit v0.3
Mostly complete, enough, I think, for Jack and Gethin to work on. Can't
get AABB,AABB to work for the life of me however, so I can't delete on
collision with an ant. Anyone able to assist with that?

Includes:

- AntEater class and source to draw a large square
- Listeners to move with setPosition on arrow keys
- 2 AABB,Circle radii - inner and outer. Respond with "flee" and "flee
harder". Turn on by un commenting draw to see the radii.