0.92 Fixed Teleporting Food
The issue we discovered with the food teleporting should be gone.
through my testing, i haven't seen it occur anymore.
An issue we have now is how the food gets dropped off. before, i had it
so it would just get the vector between the foods location and the ant
hills location and it will just go straight there. now with obstacles,
im not sure how it'll work. Cus the ant will try and go through
obstacles and when the ant is on the obstacle, it will crash. Maybe A*
but I don't know