Anteater Fix plus possible anthill config
Not sure what to do about the anthill. As an ant eater would just sit
around the ant hill until ants to come out. At the moment I've got it so
the ant eater should go the other direction when it gets near the
anthill but sometimes this doesn't work. Also it can't see ants when its
there. Done a fix for the border for anteater.